discord/requirements.txt
aiohttp>=3.7.4
alive-progress==1.6.2
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
discord-py-slash-command==2.3.1
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-secret-manager==2.5.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
idna==2.10
libcst==0.3.19
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
schedule==1.1.0
six==1.16.0
spotipy==2.18.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
yarl==1.6.3
